% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coach_module.R
\name{coach_ui}
\alias{coach_ui}
\title{BisonExplorR Coach user interface}
\usage{
coach_ui(id, title = "BisonExplorR Coach")
}
\arguments{
\item{id}{Shiny module ID.}

\item{title}{Heading displayed above the coach.}
}
\value{
A Shiny UI tag list.
}
\description{
BisonExplorR Coach user interface
}
